+++ recipes/default.rb
@@ -1,8 +1,8 @@
 #
-# Cookbook Name:: numad-nativex
+# Cookbook Name:: numad-simplyadrian
 # Recipe:: default
 #
-# Copyright 2014, NativeX
+# Copyright 2014, simplyadrian
 #
 # All rights reserved - Do Not Redistribute
 #